Refactor renderer registries
