+++ src/main/java/com/ultreon/mods/pixelguns/PixelGunsClient.java
@@ -1,11 +1,8 @@
 package com.ultreon.mods.pixelguns;
 
-import com.ultreon.mods.pixelguns.client.armor.renderer.ArmoredArmorRenderer;
-import com.ultreon.mods.pixelguns.client.armor.renderer.HazardArmorRenderer;
 import com.ultreon.mods.pixelguns.client.gui.ingame.WorkshopScreen;
-import com.ultreon.mods.pixelguns.client.item.renderer.*;
 import com.ultreon.mods.pixelguns.registry.*;
-import com.ultreon.mods.pixelguns.util.ModelPredicateProvider;
+import com.ultreon.mods.pixelguns.registry.ModelPredicateRegistry;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -13,26 +10,25 @@ import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.client.gui.screen.ingame.HandledScreens;
 import net.minecraft.client.render.RenderLayer;
-import software.bernie.geckolib3.renderers.geo.GeoArmorRenderer;
-import software.bernie.geckolib3.renderers.geo.GeoItemRenderer;
 
 @Environment(value = EnvType.CLIENT)
 public class PixelGunsClient implements ClientModInitializer {
 
 
     public void onInitializeClient() {
+        KeybindRegistry.registerKeybinds();
+
         BlockRenderLayerMap.INSTANCE.putBlock(BlockRegistry.WORKSHOP, RenderLayer.getCutout());
 
-        KeybindRegistry.registerKeybinds();
+
 
         HandledScreens.register(ScreenHandlerRegistry.WORKSHOP_SCREEN_HANDLER, WorkshopScreen::new);
 
-        ModelPredicateProvider.registerModels();
+        ModelPredicateRegistry.registerModelPredicates();
 
-        EntityRegistry.registerRenderers();
-
-        GeoArmorRenderer.registerArmorRenderer(new ArmoredArmorRenderer(), ItemRegistry.ARMORED_VEST);
-        GeoArmorRenderer.registerArmorRenderer(new HazardArmorRenderer(), ItemRegistry.GAS_MASK);
+        RendererRegistry.registerEntityRenderers();
+        RendererRegistry.registerItemRenderers();
+        RendererRegistry.registerArmorRenderers();
 
         ClientPlayNetworking.registerGlobalReceiver(PixelGuns.RECOIL_PACKET_ID, (client, handler, buf, sender) -> {
             float kick = buf.readFloat();
@@ -45,10 +41,6 @@ public class PixelGunsClient implements ClientModInitializer {
             });
         });
         
-        GeoItemRenderer.registerItemRenderer(ItemRegistry.INFINITY_GUN, new InfinityGunItemRenderer());
-        GeoItemRenderer.registerItemRenderer(ItemRegistry.ROCKET_LAUNCHER, new RocketLauncherItemRenderer());
-        GeoItemRenderer.registerItemRenderer(ItemRegistry.KATANA, new KatanaItemRenderer());
-        GeoItemRenderer.registerItemRenderer(ItemRegistry.CROWBAR, new CrowbarItemRenderer());
-        GeoItemRenderer.registerItemRenderer(ItemRegistry.GRENADE, new GrenadeItemRenderer());
+
     }
 }
